requirements.txt
discord-py-interactions==5.13.2
requests==2.32.3